requirements.txt
pyspark
pandas
numpy
google-cloud-bigquery
google-cloud-storage
google-auth
requests
python-dotenv
pandas-gbq
apache-airflow
apache-airflow-providers-google
pytz
tenacity
urllib3
pytest
black
flake8
gunicorn
py4j
kagglehub
kaggle
confluent-kafka

